pyproject.toml
[tool.poetry]
name = "hexocean-task"
version = "0.1.0"
description = ""
authors = ["Danis Idiatullin <i.danis.idiatullin@gmail.com>"]
readme = "README.md"
packages = [{include = "hexocean_task"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
pillow = "^9.4.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
